+++ src/FileAdder/FileAdder.php
@@ -18,7 +18,7 @@ use Spatie\Medialibrary\Support\File;
 use Spatie\Medialibrary\Support\RemoteFile;
 use Spatie\Medialibrary\ImageGenerators\FileTypes\Image as ImageGenerator;
 use Spatie\Medialibrary\Jobs\GenerateResponsiveImagesJob;
-use Spatie\Medialibrary\MediaCollection\MediaCollection;
+use Spatie\Medialibrary\HasMedia\MediaCollection;
 use Spatie\Medialibrary\Models\Media;
 use Symfony\Component\HttpFoundation\File\File as SymfonyFile;
 use Symfony\Component\HttpFoundation\File\UploadedFile;